KNN-BMI-main/requirements.txt
autogluon==1.2
catboost==1.2.2
imbalanced-learn==0.13.0
lightgbm==4.5.0
matplotlib==3.10.0
numpy==1.26.4
pandas==2.2.3
PyYAML==6.0.1
scikit-learn==1.5.2
scikit-optimize==0.10.2
scipy==1.10.1
seaborn==0.13.2
shap==0.47.1
statsmodels==0.14.4
xgboost==1.6.2
ydata-profiling==4.16.1